Added remaining default commands
resolved comments
